requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
SQLAlchemy==1.4.22
web3==5.23.1
psycopg2-binary==2.9.1
python-dotenv==0.18.0
